// frontend/src/app/models/Despesa.ts
import { Cartao } from "./Cartao";
import { Categoria } from "./Categoria";
import { Mes } from "./Mes";

export class Despesa {
  id: number;
  cartaoId: number;
  cartao: Cartao;
  descricao: string;
  categoriaId: number;
  categoria: Categoria;
  valor: number;
  dia: number;
  mesId: number;
  mes: Mes;
  ano: number;
  usuarioId: number;
}